requirements.txt
moviepy
openai-whisper
pydub